genconsts: Fixed L suffix for large numbers
The check so far only recognixed the suffix "L" if also the prefix
"0x" was used. This does not necessarily needs to be true.

The new check either requires the "0x" prefix or the value without
the "L" suffix needs to be numeric.

Otherwise a number like "19392L" would have been rendered as
"long(19392L)" in python, which did not compile.

Also fixing the order of imports in the rendered file so that the
imports are at the very beginning.